requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0
python-multipart==0.0.6

# LangChain Ecosystem
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.19

# Vector Store and Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.3.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pgvector==0.2.4

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
markdown==3.5.2
unstructured==0.12.2

# OpenAI
openai==1.10.0

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
aiofiles==23.2.1

# Monitoring and Logging
python-json-logger==2.0.7

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
